+++ src/Checkout/ExpressCheckout/ExpressCheckoutButtonData.php
@@ -51,6 +51,8 @@ class ExpressCheckoutButtonData extends Struct
 
     protected string $cancelRedirectUrl;
 
+    protected bool $disablePayLater;
+
     public function getProductDetailEnabled(): bool
     {
         return $this->productDetailEnabled;
@@ -154,5 +156,15 @@ class ExpressCheckoutButtonData extends Struct
     public function getCancelRedirectUrl(): string
     {
         return $this->cancelRedirectUrl;
+    }
+
+    public function isDisablePayLater(): bool
+    {
+        return $this->disablePayLater;
+    }
+
+    public function setDisablePayLater(bool $disablePayLater): void
+    {
+        $this->disablePayLater = $disablePayLater;
     }
 }