+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-checkout/index.js
@@ -134,6 +134,8 @@ Component.register('swag-paypal-checkout', {
         paymentMethodCriteria() {
             const criteria = new Criteria(1, 500);
 
+            criteria.addAssociation('media');
+
             criteria.addFilter(Criteria.equals('plugin.name', 'SwagPayPal'));
             criteria.addSorting(Criteria.sort('position', 'ASC'), true);