+++ src/Resources/app/storefront/src/checkout/swag-paypal.pay-later.js
@@ -1,0 +1,7 @@
+import SwagPaypalAbstractStandalone from './swag-paypal.abstract-standalone';
+
+export default class SwagPaypalSepa extends SwagPaypalAbstractStandalone {
+    getFundingSource(paypal) {
+        return paypal.FUNDING.PAYLATER;
+    }
+}