+++ src/Setting/Settings.php
@@ -43,12 +43,16 @@ final class Settings
     public const SPB_BUTTON_COLOR = self::SYSTEM_CONFIG_DOMAIN . 'spbButtonColor';
     public const SPB_BUTTON_SHAPE = self::SYSTEM_CONFIG_DOMAIN . 'spbButtonShape';
     public const SPB_BUTTON_LANGUAGE_ISO = self::SYSTEM_CONFIG_DOMAIN . 'spbButtonLanguageIso';
-    public const SPB_SHOW_PAY_LATER = self::SYSTEM_CONFIG_DOMAIN . 'spbShowPayLater';
     public const PUI_CUSTOMER_SERVICE_INSTRUCTIONS = self::SYSTEM_CONFIG_DOMAIN . 'puiCustomerServiceInstructions';
     public const INSTALLMENT_BANNER_ENABLED = self::SYSTEM_CONFIG_DOMAIN . 'installmentBannerEnabled';
     public const LOGGING_LEVEL = self::SYSTEM_CONFIG_DOMAIN . 'loggingLevel';
     public const EXCLUDED_PRODUCT_IDS = self::SYSTEM_CONFIG_DOMAIN . 'excludedProductIds';
     public const EXCLUDED_PRODUCT_STREAM_IDS = self::SYSTEM_CONFIG_DOMAIN . 'excludedProductStreamIds';
+
+    /**
+     * @deprecated tag:v6.0.0 - Will be removed without replacement
+     */
+    public const SPB_SHOW_PAY_LATER = self::SYSTEM_CONFIG_DOMAIN . 'spbShowPayLater';
 
     /**
      * @deprecated tag:v6.0.0 - Will be removed without replacement.
@@ -92,7 +96,7 @@ final class Settings
         self::SPB_ALTERNATIVE_PAYMENT_METHODS_ENABLED => false,
         self::SPB_BUTTON_COLOR => 'gold',
         self::SPB_BUTTON_SHAPE => 'rect',
-        self::SPB_SHOW_PAY_LATER => true,
+        self::SPB_SHOW_PAY_LATER => false,
         self::PLUS_CHECKOUT_ENABLED => false,
         self::INSTALLMENT_BANNER_ENABLED => true,
         self::LOGGING_LEVEL => Logger::WARNING,
@@ -101,8 +105,19 @@ final class Settings
         self::EXCLUDED_PRODUCT_STREAM_IDS => [],
     ];
 
+    /**
+     * @deprecated tag:v6.0.0 - Will be removed without replacement.
+     */
     public const MERCHANT_LOCATION_GERMANY = 'germany';
+
+    /**
+     * @deprecated tag:v6.0.0 - Will be removed without replacement.
+     */
     public const MERCHANT_LOCATION_OTHER = 'other';
+
+    /**
+     * @deprecated tag:v6.0.0 - Will be removed without replacement.
+     */
     public const VALID_MERCHANT_LOCATIONS = [
         self::MERCHANT_LOCATION_GERMANY,
         self::MERCHANT_LOCATION_OTHER,